+++ acc-importer.php
@@ -8,7 +8,7 @@
  * Plugin Name:       ACC User Importer
  * Plugin URI:        https://github.com/acc-wp/acc_user_importer
  * Description:       A plugin for synchronizing users from the <a href="http://alpineclubofcanada.ca">Alpine Club of Canada</a> national office.
- * Version:           4.0.4
+ * Version:           4.0.5
  * Author:            Francois Bessette, Claude Vessaz, Raz Peel, Karine Frenette Gaufre, Keith Dunwoody
  * License:           GPL-2.0+
  * License URI:       http://www.gnu.org/licenses/gpl-2.0.txt
@@ -40,7 +40,7 @@ const ACCUM_SEC = "accUM_sec_"; //The per-section settings
 /**
  * Current plugin version.
  */
-define("ACC_USER_IMPORTER_VERSION", "4.0.4");
+define("ACC_USER_IMPORTER_VERSION", "4.0.5");
 
 /**
  * Plugin deactivation.